more namespace changes
