+++ src/Cassandra.Tests/Connections/Control/ResolvedContactPointTests.cs
@@ -18,6 +18,7 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Cassandra.Connections;
+using Cassandra.Connections.Control;
 using NUnit.Framework;
 
 namespace Cassandra.Tests.Connections.Control
