+++ src/Cassandra.Tests/Connections/SniContactPointAndResolverTests.cs
@@ -22,6 +22,7 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Cassandra.Connections;
+using Cassandra.Connections.Control;
 using Cassandra.Helpers;
 using Moq;
 using NUnit.Framework;
