+++ src/Cassandra.Tests/Connections/TestHelpers/FakeConnectionEndPoint.cs
@@ -16,8 +16,9 @@
 using System.Net;
 using System.Threading.Tasks;
 using Cassandra.Connections;
+using Cassandra.Connections.Control;
 
-namespace Cassandra.Tests.Connections.Control.TestHelpers
+namespace Cassandra.Tests.Connections.TestHelpers
 {
     internal class FakeConnectionEndPoint : IConnectionEndPoint
     {
