+++ src/Cassandra.Tests/Connections/TestHelpers/FakeProtocolVersionNegotiator.cs
@@ -19,7 +19,7 @@ using Cassandra.Connections;
 using Cassandra.Connections.Control;
 using Cassandra.Serialization;
 
-namespace Cassandra.Tests.Connections.Control.TestHelpers
+namespace Cassandra.Tests.Connections.TestHelpers
 {
     internal class FakeProtocolVersionNegotiator : IProtocolVersionNegotiator
     {
