+++ src/Cassandra.Tests/Connections/TestHelpers/FakeServerEventsSubscriber.cs
@@ -18,7 +18,7 @@ using System.Threading.Tasks;
 using Cassandra.Connections;
 using Cassandra.Connections.Control;
 
-namespace Cassandra.Tests.Connections.Control.TestHelpers
+namespace Cassandra.Tests.Connections.TestHelpers
 {
     internal class FakeServerEventsSubscriber : IServerEventsSubscriber
     {
