+++ src/Cassandra.Tests/Requests/PrepareHandlerTests.cs
@@ -27,7 +27,7 @@ using Cassandra.Requests;
 using Cassandra.Responses;
 using Cassandra.Serialization;
 using Cassandra.SessionManagement;
-using Cassandra.Tests.Connections.Control.TestHelpers;
+using Cassandra.Tests.Connections.TestHelpers;
 using Moq;
 
 using NUnit.Framework;
