+++ src/Cassandra/Connections/Control/IContactPoint.cs
@@ -12,16 +12,20 @@
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
+// 
 
-using Cassandra.Connections.Control;
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
-namespace Cassandra.Tests.Connections.Control.TestHelpers
+namespace Cassandra.Connections.Control
 {
-    internal class FakeSupportedOptionsInitializerFactory : ISupportedOptionsInitializerFactory
+    internal interface IContactPoint : IEquatable<IContactPoint>
     {
-        public ISupportedOptionsInitializer Create(Metadata metadata)
-        {
-            return new FakeSupportedOptionsInitializer(metadata);
-        }
+        bool CanBeResolved { get; }
+
+        string StringRepresentation { get; }
+
+        Task<IEnumerable<IConnectionEndPoint>> GetConnectionEndPointsAsync(bool refreshCache);
     }
 }
