+++ src/Cassandra/Connections/HostnameContactPoint.cs
@@ -19,6 +19,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
+using Cassandra.Connections.Control;
 
 namespace Cassandra.Connections
 {
