+++ src/Cassandra/Connections/IConnectionEndpoint.cs
@@ -18,6 +18,7 @@ using System;
 using System.Net;
 using System.Net.Security;
 using System.Threading.Tasks;
+using Cassandra.Connections.Control;
 
 namespace Cassandra.Connections
 {
