+++ src/Cassandra/Host.cs
@@ -19,6 +19,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using Cassandra.Connections;
+using Cassandra.Connections.Control;
 using Cassandra.SessionManagement;
 
 namespace Cassandra
